Atm (atms-bank v0.1.0)

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

deposit(process_name, amount)

Specs

deposit(atom(), float()) :: :ok

Deposits money for the provided logged-in account for the Atm GenServer process

Examples

iex> Atm.deposit(:atm1, 1250.54)
:ok
◉ ◉ ◉ "Deposited $1250.54 succesfully" ◉ ◉ ◉

get_balance(process_name)

Specs

get_balance(atom()) :: :ok

Get the balance for the provided logged-in account for the Atm GenServer process

Examples

iex> Atm.get_balance(:atm1)
:ok
◉ ◉ ◉ "Your account balance is 8244.22" ◉ ◉ ◉

get_transactions(process_name)

Specs

get_transactions(atom()) :: :ok

Get the deposits and withdrawalsfor the provided logged-in account for the Atm GenServer process

Examples

iex> Atm.deposit(:atm1, 1250.54)
:ok
◉ ◉ ◉ "Deposits [1250.54]" ◉ ◉ ◉
◉ ◉ ◉ "Withdrawals []" ◉ ◉ ◉

login(process_name, input)

Specs

login(atom(), %{card_digits: String, pin: String}) :: :ok

Login with the specified card digits matching the bank bin with the provided GenServer named process and pin

Examples

iex> Atm.login(:atm1, %{card_digits: "4101-7712-3456-7890", pin: "0021"})
:ok
✓ ✓ ✓ Succesful login ✓ ✓ ✓

iex> Atm.login(:atm1, %{card_digits: "4101-7712-3456-7812", pin: "0021"})
✕ ✕ ✕ Account does not exist ✕ ✕ ✕
:ok

iex> Atm.login(:atm1, %{card_digits: "4101-7712-3456-7890", pin: "1235"})
✕ ✕ ✕ Invalid PIN ✕ ✕ ✕
:ok

logout(process_name)

Specs

logout(atom()) :: :ok

Logout from the existing session for the provided Atm GenServer process.

Examples

iex> Atm.logout(:atm1)
✌ ✌ ✌ Succesful logout ✌ ✌ ✌
:ok

start_link(atm_name)

Specs

start_link(atom()) :: {:ok, pid()} | {:already_started, pid()}

Starts an Atm GenServer supervised process with a given atom process name

Examples

iex> Atm.start_link(:atm1)
{:ok, #PID<0.198.0>}

iex> Atm.start_link(:atm1)
{:already_started, #PID<0.198.0>}

withdraw(process_name, amount)

Specs

withdraw(atom(), float()) :: :ok

Withdraws money for the provided logged-in account for the Atm GenServer process

Examples

iex> Atm.withdraw(:atm1, 8000)
:ok
◉ ◉ ◉ "Please take $: 8000 from the drawer 

Balance $: 244.21999999999935" ◉ ◉ ◉

iex> Atm.withdraw(:atm1, 8000)
✕ ✕ ✕ Can't exceed daily withdrawal total. Try with a smaller withdraw amount ✕ ✕ ✕
:ok

iex> Atm.withdraw(:atm1, 80001)
✕ ✕ ✕ Can't withdraw more than current balance ✕ ✕ ✕
:ok